+++ src/SFA.DAS.ApprenticeCommitments.Api.AcceptanceTests/Features/GetApprenticeshipsSteps.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Newtonsoft.Json;
 using SFA.DAS.ApprenticeCommitments.Data.Models;
-using SFA.DAS.ApprenticeCommitments.Models;
+using SFA.DAS.ApprenticeCommitments.DTOs;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
