+++ src/SFA.DAS.ApprenticeCommitments/Application/Commands/VerifyRegistrationCommand/VerifyRegistrationCommandHandler.cs
@@ -2,7 +2,7 @@
 using SFA.DAS.ApprenticeCommitments.Data;
 using SFA.DAS.ApprenticeCommitments.Data.Models;
 using SFA.DAS.ApprenticeCommitments.Exceptions;
-using SFA.DAS.ApprenticeCommitments.Models;
+using SFA.DAS.ApprenticeCommitments.DTOs;
 using System.Net.Mail;
 using System.Threading;
 using System.Threading.Tasks;
