+++ src/SFA.DAS.ApprenticeCommitments/DTOs/ApprenticeDto.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net.Mail;
 
-namespace SFA.DAS.ApprenticeCommitments.Models
+namespace SFA.DAS.ApprenticeCommitments.DTOs
 {
     public class ApprenticeDto
     {
