+++ src/SFA.DAS.ApprenticeCommitments/DTOs/RegistrationDto.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace SFA.DAS.ApprenticeCommitments.Models
+namespace SFA.DAS.ApprenticeCommitments.DTOs
 {
     public class RegistrationDto
     {
